backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
openai==1.35.3
python-dotenv==1.0.1
httpx==0.27.0
pydantic==2.12.5
pydantic-core==2.41.5
